+++ src/components/domain/ProductPlaceHolder/index.stories.tsx
