+++ src/components/domain/index.ts
@@ -1,0 +1,3 @@
+export { default as Product } from './Product';
+export { default as ProductCard } from './ProductCard';
+export { default as ProductPlaceHolder } from './ProductPlaceHolder';
